Explain example data set member ordered-by system list
The example data set "members" list is ordered-by system and thus
results are in arbitrary order without sort-by. Add a note in the
Example Queries section about this.

Furthermore the result-set in ex-where-array-match-on-attribute returned
a result-set in another order then the example data set order, which
might have proven the point of ordered-by system list but it is
confusing. However, since the fix for #23 (where match on attribute
email-address) this result-set is no longer in a different order because
the out of order nodes are omitted due to the where clause,

Fixes #24